refactor(frontend): move theme definition to a separate file
