+++ frontend/src/index.tsx
@@ -12,69 +12,16 @@
  *******************************************************************************/
 import { ApolloProvider } from "@apollo/client";
 import CssBaseline from "@material-ui/core/CssBaseline";
-import { createTheme, ThemeProvider } from "@material-ui/core/styles";
+import { ThemeProvider } from "@material-ui/core/styles";
 import { ApolloGraphQLClient } from "ApolloGraphQLClient";
 import { Main } from "main/Main";
 import ReactDOM from "react-dom";
-import { theme } from "@eclipse-sirius/sirius-components";
+import { siriusWebTheme } from "theme";
 import { BrowserRouter } from "react-router-dom";
 import "./fonts.css";
 import "./reset.css";
 import "./Sprotty.css";
 import "./variables.css";
-
-const baseTheme = createTheme({
-  ...theme,
-  palette: {
-    type: "light",
-    primary: {
-      main: "#BE1A78",
-      dark: "#851254",
-      light: "#CB4793",
-    },
-    secondary: {
-      main: "#261E58",
-      dark: "#1A153D",
-      light: "#514B79",
-    },
-    text: {
-      primary: "#261E58",
-      disabled: "#B3BFC5",
-      hint: "#B3BFC5",
-    },
-    error: {
-      main: "#DE1000",
-      dark: "#9B0B00",
-      light: "#E43F33",
-    },
-    divider: "#B3BFC5",
-  },
-  props: {
-    MuiAppBar: {
-      color: "secondary",
-    },
-  },
-  overrides: {
-    MuiSnackbarContent: {
-      root: {
-        backgroundColor: "#7269A4",
-      },
-    },
-  },
-});
-
-const siriusWebTheme = createTheme(
-  {
-    overrides: {
-      MuiAvatar: {
-        colorDefault: {
-          backgroundColor: baseTheme.palette.primary.main,
-        },
-      },
-    },
-  },
-  baseTheme
-);
 
 const style = {
   display: "grid",
